Added ecs in the server

+++ client-core/src/main/java/com/ylinor/client/network/NetworkIdentifierComponent.java
@@ -3,13 +3,13 @@ package com.ylinor.client.network;
 import com.ylinor.library.util.ecs.component.Component;
 
 public final class NetworkIdentifierComponent extends Component {
-    private long identifier;
+    private int identifier;
 
-    public long getIdentifier() {
+    public int getIdentifier() {
         return identifier;
     }
 
-    public void setIdentifier(long identifier) {
+    public void setIdentifier(int identifier) {
         this.identifier = identifier;
     }
 }
